# docs/veyon/gateway-openapi.yaml
openapi: 3.0.3
info:
  title: QLPM Veyon Gateway API
  version: 1.0.0
  description: Internal API between Rails QLPM and Veyon Gateway.
servers:
  - url: http://veyon-gateway:8088
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    OpenConnectionRequest:
      type: object
      required: [host, auth_method]
      properties:
        host:
          type: string
          description: Target Windows client endpoint as host:service_port. In QLPM this is built from veyon_hosts.host and veyon_hosts.service_port, usually port 11100.
          example: lab-a-pc-01.school.local:11100
        auth_method:
          type: string
          enum: [auth_keys, auth_logon, auth_simple, auth_ldap]
        credentials:
          type: object
          additionalProperties: true
    OpenConnectionResponse:
      type: object
      required: [connection_uid, valid_until]
      properties:
        connection_uid:
          type: string
        valid_until:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
paths:
...
  /v1/connections/open:
    post:
      summary: Open or reuse authenticated connection to host
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenConnectionRequest'
      responses:
        '200':
          description: Connection opened
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenConnectionResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many connections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...